fix(GAT-0000): fixed users load (#1369)
