+++ src/app/[locale]/account/team/[teamId]/(withLeftNav)/team-management/components/AddTeamMemberRows/AddTeamMemberRows.tsx
@@ -2,36 +2,56 @@ import { useMemo } from "react";
 import { Control, FieldArrayWithId, FieldValues } from "react-hook-form";
 import { IconButton } from "@mui/material";
 import { AddTeamMember, UserAndRoles } from "@/interfaces/AddTeamMember";
+import { User } from "@/interfaces/User";
 import Box from "@/components/Box";
 import InputWrapper from "@/components/InputWrapper";
-import { SelectOptionsType } from "@/components/Select/Select";
+import useGet from "@/hooks/useGet";
+import apis from "@/config/apis";
 import {
     addTeamMemberFormFields,
     getRoleOptions,
 } from "@/config/forms/addTeamMember";
 import { AddIcon, RemoveIcon, SearchRoundedIcon } from "@/consts/icons";
+import { getAvailableUsers } from "../AddTeamMemberDialog/AddTeamMemberDialog.utils";
 
 interface AddTeamMemberRowsProps<TFieldValues extends FieldValues> {
     fields: FieldArrayWithId<AddTeamMember, "userAndRoles", "id">[];
     control: Control<TFieldValues>;
-    userOptions: SelectOptionsType[];
     userPermissions: { [key: string]: boolean };
     remove: (fieldId: number) => void;
     append: (newRow: UserAndRoles) => void;
     onInputChangeUser?: (value: string) => void;
-    isLoadingUsers?: boolean;
+    team: {
+        users: User[];
+    };
 }
 
 const AddTeamMemberRows = <TFieldValues extends FieldValues>({
     fields,
     control,
-    userOptions,
+    team,
     userPermissions,
     remove,
     append,
     onInputChangeUser,
-    isLoadingUsers = false,
 }: AddTeamMemberRowsProps<TFieldValues>) => {
+    // TODO: switch to typeahead propery
+    // - difficult due to use of useFieldArray and shared options
+    // - defaulting to full list of all users as a temporary measure..
+    // const [searchName, setSearchName] = useState("");
+    // const searchNameDebounced = useDebounce(searchName, 500);
+
+    const { data: users = [], isLoading: isLoadingUsers } = useGet<User[]>(
+        `${apis.usersV1Url}?mini`, // filterNames=${searchNameDebounced}`,
+        {
+            shouldFetch: true, // !!searchNameDebounced
+        }
+    );
+
+    const userOptions = useMemo(() => {
+        return getAvailableUsers(team.users, users);
+    }, [team, users]);
+
     const roleOptionsFiltered = useMemo(() => {
         return getRoleOptions().filter(role =>
             role.permissions?.every(permission => userPermissions[permission])
